pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "tangods_moenchzmq"
description = "processing ZMQ server for a moench detector with tango DeviceServer"
keywords = ["tango device", "tango", "pytango", "moench"]
readme = "README.md"
authors = [
    {name = "Leonid Lunin", email = "lunin.leonid@gmail.com"},
]
requires-python = ">=3.10"
dependencies = [
    'pytango >= 9.3.6',
    'numpy >= 1.24.1',
    'pyzmq >= 23.2.0',
    'numba >= 0.57',
    'nexusformat >= 1.0.2',
    'Pillow'
]
license = {text = "MIT"}
dynamic = ["version"]

[project.urls]
Homepage = "https://github.com/lrlunin/pytango-moenchZmqServer"

[project.scripts]
MoenchZmqServer = "tangods_moenchzmq:main"
MoenchZmqTestServer = "testserver_moenchzmq:main"

[tool.setuptools.dynamic]
version = {file = "VERSION"}

[tool.setuptools.packages.find]

[tool.setuptools.package-data]
tangods_moenchzmq = ["VERSION"]
"tangods_moenchzmq.default_images" = ["*.npy"]
"tangods_moenchzmq.reorder_tables" = ["*.npy"]
"testserver_moenchzmq.util" = ["*.ttf"]
